Use tm_kinematics API for calculate IK, FK
Originall, I use Eigen to calculate forward kinematics, but after
comparing with the API from tm_kinematics, I found tm_kin is much more
faster thet eigen, so I delete the module using eigen to calculate FK.
For IK, I use tm_kin to calculate to q. But something is worth notices
that if you set the q to all zeros, AKA "home" position, the inverse
result will be 0 " -180 0 180 0 0 ", so be carefull with that.